Django-Middleware-0x03/requirements.txt
asgiref==3.8.1
autopep8==2.3.2
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.1
django-filter==25.1
djangorestframework==3.16.0
dotenv==0.9.9
drf-nested-routers==0.94.2
idna==3.10
mysql-connector-python==9.3.0
parameterized==0.9.0
pillow==11.3.0
pycodestyle==2.13.0
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.3
urllib3==2.4.0
